requirements.txt
PyQt6>=6.7.0
google-genai>=1.5.0
python-dotenv>=1.0.1
python-docx>=1.1.2
reportlab>=4.2.0
openpyxl>=3.1.3
matplotlib>=3.9.0
